+++ CHANGELOG.md
@@ -28,7 +28,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 - `compass` event did not work with variables in `compass` location
 - `quest canceler` did not resolve global variables
 - `quest canceler` used `loc` instead of `location` key for teleportation
-- `journal` language only changed after rejoin when changing player language 
+- `journal` language only changed after rejoin when changing player language
+- `journal` entry renaming or deleting via command now affects all specified pointers and not only one
 ### Security
 
 ## [2.2.1] - 2025-01-12